requirements-scraper.txt
requests==2.31.0
beautifulsoup4==4.12.2
undetected-chromedriver==3.5.4
selenium==4.15.2
fake-useragent==1.4.0
flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.1.0
google-generativeai==0.8.5
numpy==2.3.1
supabase==1.2.0
psycopg2-binary==2.9.9 